[RENAME] change reviewGET filename
