+++ functions/api/routes/review/index.js
@@ -6,7 +6,7 @@ router.post("/", checkUser, require("./reviewPOST"));
 router.delete("/:id", checkUser, require("./reviewDELETE"));
 router.get("/:id", checkUser, require("./reviewDetailGET"));
 router.get("/tag", checkUser, require("./reviewTagListGET"));
-router.get("/major/:majorId/writer/:writerFilter", checkUser, require("./reviewListGET"));
+router.get("/major/:majorId/writer/:writerFilter", checkUser, require("./reviewMajorListGET"));
 router.put("/:id", checkUser, require("./reviewPUT"));
 
 module.exports = router;
